requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
abseil-cpp=20211102.0=h0e60522_0
altair=5.2.0=pyhd8ed1ab_0
anyio=4.1.0=pyhd8ed1ab_0
argon2-cffi=23.1.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py312h2bbff1b_0
arrow=1.3.0=pyhd8ed1ab_0
arrow-cpp=11.0.0=ha81ea56_2
asttokens=2.4.1=pyhd8ed1ab_0
async-lru=2.0.4=pyhd8ed1ab_0
attrs=23.1.0=pyh71513ae_1
aws-c-common=0.6.8=h2bbff1b_1
aws-c-event-stream=0.1.6=hd77b12b_6
aws-checksums=0.1.11=h2bbff1b_2
aws-sdk-cpp=1.8.185=hd77b12b_1
babel=2.13.1=pyhd8ed1ab_0
beautifulsoup4=4.12.2=pyha770c72_0
blas=1.0=mkl
bleach=6.1.0=pyhd8ed1ab_0
blinker=1.7.0=pyhd8ed1ab_0
boost-cpp=1.78.0=h5b4e17d_0
bottleneck=1.3.5=py312he558020_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
bzip2=1.0.8=he774522_0
c-ares=1.19.1=h2bbff1b_0
ca-certificates=2023.11.17=h56e8100_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.3.2=pyhd8ed1ab_0
certifi=2023.11.17=pyhd8ed1ab_0
cffi=1.16.0=py312h2bbff1b_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=win_pyh7428d3b_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.2.0=py312h59b6b97_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.7=py312hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
expat=2.5.0=hd77b12b_0
fonttools=4.25.0=pyhd3eb1b0_0
fqdn=1.5.1=pyhd8ed1ab_0
freetype=2.12.1=ha860e81_0
gflags=2.2.2=ha925a31_1004
giflib=5.2.1=h8cc25b3_3
gitdb=4.0.11=pyhd8ed1ab_0
gitpython=3.1.40=pyhd8ed1ab_0
glog=0.5.0=h4797de2_0
grpc-cpp=1.48.2=hfe90ff0_1
idna=3.6=pyhd8ed1ab_0
importlib-metadata=6.10.0=pyha770c72_0
importlib_metadata=6.10.0=hd8ed1ab_0
importlib_resources=6.1.1=pyhd8ed1ab_0
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.26.0=pyha63f2e9_0
ipython=8.18.1=pyh7428d3b_3
ipywidgets=8.1.1=pyhd8ed1ab_0
isoduration=20.11.0=pyhd8ed1ab_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
jpeg=9e=h2bbff1b_1
json5=0.9.14=pyhd8ed1ab_0
jsonpointer=2.0=py_0
jsonschema=4.20.0=pyhd8ed1ab_0
jsonschema-specifications=2023.11.2=pyhd8ed1ab_0
jsonschema-with-format-nongpl=4.20.0=pyhd8ed1ab_0
jupyter-lsp=2.2.1=pyhd8ed1ab_0
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=5.5.0=py312haa95532_0
jupyter_events=0.9.0=pyhd8ed1ab_0
jupyter_server=2.12.1=pyhd8ed1ab_0
jupyter_server_terminals=0.4.4=pyhd8ed1ab_1
jupyterlab=4.0.9=pyhd8ed1ab_0
jupyterlab_pygments=0.3.0=pyhd8ed1ab_0
jupyterlab_server=2.25.2=pyhd8ed1ab_0
jupyterlab_widgets=3.0.9=pyhd8ed1ab_0
kiwisolver=1.4.4=py312hd77b12b_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libcurl=8.4.0=h86230a5_1
libdeflate=1.17=h2bbff1b_1
libevent=2.1.10=h91dae50_4
libffi=3.4.4=hd77b12b_0
libpng=1.6.39=h8cc25b3_0
libprotobuf=3.20.3=h23ce68f_0
libsodium=1.0.18=h8d14728_1
libssh2=1.10.0=he2ea4bf_2
libthrift=0.15.0=h4364b78_2
libtiff=4.5.1=hd77b12b_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
markdown-it-py=3.0.0=pyhd8ed1ab_0
markupsafe=2.1.1=py312h2bbff1b_0
matplotlib-base=3.8.0=py312hc7c4135_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.0=pyhd8ed1ab_0
mistune=3.0.2=pyhd8ed1ab_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py312h2bbff1b_1
mkl_fft=1.3.8=py312h2bbff1b_0
mkl_random=1.2.4=py312h59b6b97_0
mlb-statsapi=1.7=pypi_0
munkres=1.1.4=py_0
nbclient=0.8.0=pyhd8ed1ab_0
nbconvert-core=7.12.0=pyhd8ed1ab_0
nbformat=5.9.2=pyhd8ed1ab_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
notebook-shim=0.2.3=pyhd8ed1ab_0
numexpr=2.8.7=py312h96b7d27_0
numpy=1.26.2=py312hfd52020_0
numpy-base=1.26.2=py312h4dde369_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.0.12=h2bbff1b_0
orc=1.7.4=h623e30f_1
overrides=7.4.0=pyhd8ed1ab_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.1=py312hc7c4135_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.0.1=py312h045eedc_0
pip=23.3.1=py312haa95532_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_1
platformdirs=4.1.0=pyhd8ed1ab_0
prometheus_client=0.19.0=pyhd8ed1ab_0
prompt-toolkit=3.0.41=pyha770c72_0
protobuf=3.20.3=py312hd77b12b_0
psutil=5.9.0=py312h2bbff1b_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=11.0.0=py312had8a6e9_2
pycparser=2.21=pyhd8ed1ab_0
pydeck=0.8.0=pyhd8ed1ab_0
pygments=2.17.2=pyhd8ed1ab_0
pyparsing=3.0.9=py312haa95532_0
python=3.12.0=h1d929f7_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.19.0=pyhd8ed1ab_0
python-json-logger=2.0.7=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2023.3.post1=py312haa95532_0
pywin32=305=py312h2bbff1b_0
pywinpty=2.0.10=py312h5da7b33_0
pyyaml=6.0.1=py312h2bbff1b_0
pyzmq=25.1.0=py312hd77b12b_0
re2=2022.04.01=h0e60522_0
referencing=0.32.0=pyhd8ed1ab_0
requests=2.31.0=pyhd8ed1ab_0
rfc3339-validator=0.1.4=pyhd8ed1ab_0
rfc3986-validator=0.1.1=pyh9f0ad1d_0
rich=13.7.0=pyhd8ed1ab_0
rpds-py=0.10.6=py312h062c2fa_0
seaborn=0.12.2=py312haa95532_0
send2trash=1.8.2=pyh08f2357_0
setuptools=68.0.0=py312haa95532_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pyhd8ed1ab_0
snappy=1.1.9=h6c2663c_0
sniffio=1.3.0=pyhd8ed1ab_0
soupsieve=2.5=pyhd8ed1ab_1
sqlite=3.41.2=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
streamlit=1.29.0=pyhd8ed1ab_0
tbb=2021.8.0=h59b6b97_0
tenacity=8.2.3=pyhd8ed1ab_0
terminado=0.18.0=pyh5737063_0
tinycss2=1.2.1=pyhd8ed1ab_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd8ed1ab_0
tomli=2.0.1=pyhd8ed1ab_0
toolz=0.12.0=pyhd8ed1ab_0
tornado=6.3.3=py312h2bbff1b_0
traitlets=5.14.0=pyhd8ed1ab_0
types-python-dateutil=2.8.19.14=pyhd8ed1ab_0
typing-extensions=4.9.0=hd8ed1ab_0
typing_extensions=4.9.0=pyha770c72_0
typing_utils=0.1.0=pyhd8ed1ab_0
tzdata=2023c=h04d1e81_0
tzlocal=3.0=pyhd8ed1ab_0
uri-template=1.3.0=pyhd8ed1ab_0
urllib3=2.1.0=py312haa95532_0
utf8proc=2.6.1=h2bbff1b_0
validators=0.22.0=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=2.1.6=py312haa95532_0
wcwidth=0.2.12=pyhd8ed1ab_0
webcolors=1.13=pyhd8ed1ab_0
webencodings=0.5.1=pyhd8ed1ab_2
websocket-client=1.7.0=pyhd8ed1ab_0
wheel=0.41.2=py312haa95532_0
widgetsnbextension=4.0.9=pyhd8ed1ab_0
winpty=0.4.3=4
xz=5.4.5=h8cc25b3_0
yaml=0.2.5=h8ffe710_2
zeromq=4.3.4=h0e60522_1
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
